temporal

+++ src/worker/sabManage.ts
@@ -22,7 +22,7 @@ class SabWorker1 {
   public playerPosition = new SharedArrayBuffer(Float64Array.BYTES_PER_ELEMENT * 2)
 
   public enemy1Positions = new SharedArrayBuffer(Float64Array.BYTES_PER_ELEMENT * numberOfEnemy1double)
-  public enemy1Hps = new SharedArrayBuffer(Float64Array.BYTES_PER_ELEMENT * numberOfEnemy1)
+  public enemy1Hps = new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT * numberOfEnemy1)
 
   public autoAttack1Positions = new SharedArrayBuffer(Float64Array.BYTES_PER_ELEMENT * numberOfAutoAttack1 * 2)
   public autoAttack1ReaminTimes = new SharedArrayBuffer(Float64Array.BYTES_PER_ELEMENT * numberOfAutoAttack1)
@@ -36,7 +36,7 @@ class SabWorker1 {
   /** arrays used in main */
   public playerPositionArr = new Float64Array(this.playerPosition)
   public enemy1PositionsArr = new Float64Array(this.enemy1Positions)
-  public enemy1HpsArr = new Float64Array(this.enemy1Hps)
+  public enemy1HpsArr = new Int32Array(this.enemy1Hps)
   public autoAttack1PositionsArr = new Float64Array(this.autoAttack1Positions)
   public autoAttack1ReaminTimesArr = new Float64Array(this.autoAttack1ReaminTimes)
 }
